requirements.txt
asgiref==3.3.2
Django==3.2.15
djangorestframework==3.11.0
pytz==2019.3
PyJWT
django-cors-headers
django-rest-passwordreset
mysqlclient
python-dotenv
gunicorn